+++ packages/profile-sync-controller/src/index.ts
@@ -1,3 +1,4 @@
 export * as SDK from './sdk';
 export * as AuthenticationController from './controllers/authentication';
 export * as UserStorageController from './controllers/user-storage';
+export * as MetaMetrics from './shared/metametrics';